refactor(#2453 R2): move R1 handler tests to restart_tests.rs (LOC gate)
The two-phase barrier grew restart.rs to 778 LOC, back over the file_size_invariant
MAX_LOC (750). Move the daemon/capability-dispatch `mod tests` (the R1 fail-closed
tests) verbatim into the sibling restart_tests.rs (name contains "test" → the
invariant skips it), alongside the App-arm barrier tests already there. `use super::*`
→ `use super::super::*` (now a grandchild of `restart`). restart.rs → 399 LOC.
All 17 restart tests pass (moved R1 + App-arm barrier); mcp_handler_files_under_max_loc
+ the merge-freshness pin pass; fmt clean.


Agend-Agent: archfix-claude-opus
Agend-Task: t-20260712034222302055-19288-5
Agend-Branch: feat/app-staged-restart-r2
Agend-Issued-At: 2026-07-12T03:43:18.424715+00:00
