+++ src/mcp/handlers/restart.rs
@@ -389,386 +389,6 @@ fn phase1_gate(succ: &mut crate::bootstrap::daemon_spawn::SuccessorHandle) -> bo
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::atomic::Ordering;
-
-    /// Tests touch the process-global `RESTART_PENDING` static and env
-    /// state. Serialise via the SINGLE crate-wide
-    /// [`crate::daemon::test_env_lock`] (shared with `daemon::restart` +
-    /// `per_tick::recovery_dispatcher`) — env mutation races across all
-    /// keys, so a module-local mutex wouldn't serialise against those
-    /// other modules' env tests (#1812). Also resets `RESTART_PENDING`
-    /// after each test so the next caller sees a clean slate.
-    fn with_env_and_reset<R>(set: &[(&str, &str)], unset: &[&str], f: impl FnOnce() -> R) -> R {
-        let _guard = crate::daemon::test_env_lock()
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let all_keys: Vec<&str> = set
-            .iter()
-            .map(|(k, _)| *k)
-            .chain(unset.iter().copied())
-            .collect();
-        let prior: Vec<(String, Option<String>)> = all_keys
-            .iter()
-            .map(|k| (k.to_string(), std::env::var(k).ok()))
-            .collect();
-        let prior_restart_pending = crate::daemon::RESTART_PENDING.load(Ordering::Acquire);
-
-        // SAFETY: test-only mutation, serialised via the mutex above.
-        unsafe {
-            for k in unset {
-                std::env::remove_var(k);
-            }
-            for (k, v) in set {
-                std::env::set_var(k, v);
-            }
-        }
-        crate::daemon::RESTART_PENDING.store(false, Ordering::Release);
-
-        let result = f();
-
-        unsafe {
-            for (k, v) in &prior {
-                match v {
-                    Some(val) => std::env::set_var(k, val),
-                    None => std::env::remove_var(k),
-                }
-            }
-        }
-        crate::daemon::RESTART_PENDING.store(prior_restart_pending, Ordering::Release);
-
-        result
-    }
-
-    /// Contract: when no supervisor signal env is set,
-    /// `handle_restart_daemon` must fail-closed — return `ok: false`
-    /// with an actionable error, and must NOT set `RESTART_PENDING`
-    /// (which would trigger `exit(42)` on the next supervisor tick
-    /// and leave the daemon dead with nothing to respawn it).
-    #[test]
-    fn handle_restart_daemon_fails_closed_when_unsupervised() {
-        with_env_and_reset(
-            // #1814 Stage 4: self-respawn is now the DEFAULT, so the legacy
-            // fail-closed (is_restart_supervised) branch is only reached via the
-            // explicit opt-out `AGEND_RESTART_HANDOFF=0`. Set it so this test
-            // exercises the legacy supervisor-required path (was: unset, which
-            // pre-Stage-4 meant legacy-by-default but now means self-respawn).
-            &[("AGEND_RESTART_HANDOFF", "0")],
-            &[
-                "AGEND_WRAPPED",
-                "AGEND_SUPERVISED",
-                "INVOCATION_ID",
-                "XPC_SERVICE_NAME",
-            ],
-            || {
-                // #2453: inject the Daemon capability so this test drives the
-                // daemon strategy — which reaches the legacy supervisor-detection
-                // branch under test — with no process-global host state.
-                // Manual tempdir — matches the std::env::temp_dir pattern
-                // used elsewhere in this crate (e.g. claim_verifier.rs
-                // tests). Avoids adding tempfile as a dev-dep just for
-                // this one test.
-                let tmp = std::env::temp_dir().join(format!(
-                    "agend-restart-test-{}-{}",
-                    std::process::id(),
-                    std::time::SystemTime::now()
-                        .duration_since(std::time::UNIX_EPOCH)
-                        .map(|d| d.as_nanos())
-                        .unwrap_or(0),
-                ));
-                std::fs::create_dir_all(&tmp).expect("create tempdir");
-                let response = handle_restart_daemon(
-                    &tmp,
-                    Some(crate::api::RestartCapability::Daemon),
-                    None,
-                    None,
-                );
-                assert_eq!(
-                    response["ok"], false,
-                    "unsupervised invocation must return ok:false, got {response}"
-                );
-                let error = response["error"].as_str().unwrap_or("");
-                assert!(
-                    error.contains("supervisor"),
-                    "error string must name the missing supervisor — got {error:?}"
-                );
-                assert!(
-                    !crate::daemon::RESTART_PENDING.load(Ordering::Acquire),
-                    "RESTART_PENDING must NOT be set when fail-closed — \
-                     otherwise the next supervisor tick will exit(42) \
-                     and leave the daemon dead with no respawn"
-                );
-                // restart-requested marker file must also NOT be written
-                // (it's both a side-effect and an integration signal).
-                assert!(
-                    !tmp.join("restart-requested").exists(),
-                    "restart-requested marker file must NOT exist when fail-closed"
-                );
-                let _ = std::fs::remove_dir_all(&tmp);
-            },
-        );
-    }
-
-    /// #1812 regression — bare GUI launch on macOS. `XPC_SERVICE_NAME` is
-    /// ambient in a macOS GUI login session (set on EVERY process,
-    /// including a bare `agend-terminal start` from Terminal.app), so the
-    /// pre-#1812 detector treated an UNsupervised daemon as supervised and
-    /// `restart_daemon` would exit(42) with nobody to respawn it (#851
-    /// defeat). With XPC dropped, the REAL handler must fail-closed: only
-    /// `XPC_SERVICE_NAME` set, all our markers unset → `ok: false` and no
-    /// `RESTART_PENDING`. Drives the production `handle_restart_daemon`, not
-    /// the bare detector.
-    #[test]
-    fn handle_restart_daemon_fails_closed_on_bare_gui_xpc_only() {
-        with_env_and_reset(
-            // #1814 Stage 4: opt into the legacy supervisor-detection path
-            // (=0) so this exercises the is_restart_supervised XPC false-positive
-            // guard — under the new self-respawn default it would otherwise hit
-            // the spawn path (which also returns ok:false, masking the intent).
-            &[("XPC_SERVICE_NAME", "0"), ("AGEND_RESTART_HANDOFF", "0")],
-            &["AGEND_WRAPPED", "AGEND_SUPERVISED", "INVOCATION_ID"],
-            || {
-                // #2453: inject the Daemon capability so the legacy XPC
-                // false-positive guard under test is reached (no host global).
-                let tmp = std::env::temp_dir().join(format!(
-                    "agend-restart-gui-test-{}-{}",
-                    std::process::id(),
-                    std::time::SystemTime::now()
-                        .duration_since(std::time::UNIX_EPOCH)
-                        .map(|d| d.as_nanos())
-                        .unwrap_or(0),
-                ));
-                std::fs::create_dir_all(&tmp).expect("create tempdir");
-                let response = handle_restart_daemon(
-                    &tmp,
-                    Some(crate::api::RestartCapability::Daemon),
-                    None,
-                    None,
-                );
-                assert_eq!(
-                    response["ok"], false,
-                    "ambient XPC_SERVICE_NAME alone must NOT be accepted as a \
-                     supervisor (the #851/#1812 macOS-GUI false-positive) — got {response}"
-                );
-                assert!(
-                    !crate::daemon::RESTART_PENDING.load(Ordering::Acquire),
-                    "RESTART_PENDING must stay unset on the bare-GUI fail-closed path"
-                );
-                assert!(
-                    !tmp.join("restart-requested").exists(),
-                    "restart-requested marker must NOT be written on fail-closed"
-                );
-                let _ = std::fs::remove_dir_all(&tmp);
-            },
-        );
-    }
-
-    /// #2098 app/owned-mode fail-closed. `agend-terminal app` (combined
-    /// TUI+daemon, run_app) never enters run_core, so it has NO RESTART_PENDING
-    /// consumer: an in-process self-respawn would set RESTART_PENDING, brick the
-    /// api session loop (api/mod.rs), and the held-flock successor would
-    /// 30s-time-out and die — latching the brick permanently. With self-respawn
-    /// the DEFAULT (#2094: `AGEND_RESTART_HANDOFF=1`), the handler MUST fail-closed
-    /// for the App capability — BEFORE selecting any path or spawning a successor,
-    /// and WITHOUT setting RESTART_PENDING. Drives the production
-    /// `handle_restart_daemon` with the injected `RestartCapability::App` (the real
-    /// app-mode state); the §3.9 `app_mode_restart_fails_closed_no_brick_2098`
-    /// PTY test in tests/self_respawn_handoff.rs is the end-to-end reproduction.
-    #[test]
-    fn handle_restart_daemon_fails_closed_in_app_mode_no_run_core() {
-        with_env_and_reset(
-            // #2094 default: self-respawn ON — the exact brick scenario. The app
-            // guard must short-circuit BEFORE this path is even selected.
-            &[("AGEND_RESTART_HANDOFF", "1")],
-            &[
-                "AGEND_WRAPPED",
-                "AGEND_SUPERVISED",
-                "INVOCATION_ID",
-                "XPC_SERVICE_NAME",
-            ],
-            || {
-                // App host: inject the App capability (the real app-mode state)
-                // so the handler dispatches to the app fail-close arm.
-                let tmp = std::env::temp_dir().join(format!(
-                    "agend-restart-appmode-{}-{}",
-                    std::process::id(),
-                    std::time::SystemTime::now()
-                        .duration_since(std::time::UNIX_EPOCH)
-                        .map(|d| d.as_nanos())
-                        .unwrap_or(0),
-                ));
-                std::fs::create_dir_all(&tmp).expect("create tempdir");
-                let response = handle_restart_daemon(
-                    &tmp,
-                    Some(crate::api::RestartCapability::App),
-                    None,
-                    None,
-                );
-                assert_eq!(
-                    response["ok"], false,
-                    "app-mode (no run_core consumer) restart must fail-closed, got {response}"
-                );
-                let error = response["error"].as_str().unwrap_or("");
-                assert!(
-                    error.contains("app"),
-                    "fail-closed error must name `app` mode (actionable) — got {error:?}"
-                );
-                assert!(
-                    !crate::daemon::RESTART_PENDING.load(Ordering::Acquire),
-                    "RESTART_PENDING must NOT be set in app mode — with no run_core \
-                     consumer it would latch and permanently brick the control plane"
-                );
-                assert!(
-                    !tmp.join("restart-requested").exists(),
-                    "restart-requested marker must NOT be written on app-mode fail-closed"
-                );
-                let _ = std::fs::remove_dir_all(&tmp);
-            },
-        );
-    }
-
-    /// #2453 Stage R1: unique per-test tempdir (matches the std::env::temp_dir
-    /// pattern used above; avoids a tempfile dev-dep).
-    fn unique_tmp(tag: &str) -> std::path::PathBuf {
-        let tmp = std::env::temp_dir().join(format!(
-            "agend-restart-{tag}-{}-{}",
-            std::process::id(),
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .map(|d| d.as_nanos())
-                .unwrap_or(0),
-        ));
-        std::fs::create_dir_all(&tmp).expect("create tempdir");
-        tmp
-    }
-
-    /// #2453 Stage R1: an INJECTED `RestartCapability::Daemon` reaches the real
-    /// restart machinery (the daemon strategy), NOT the app fail-close. With the
-    /// legacy path forced (`AGEND_RESTART_HANDOFF=0`) and no supervisor present,
-    /// the Daemon arm returns the supervisor-required error — proving dispatch
-    /// routed to the daemon strategy (whose error names "supervisor"), not the
-    /// app arm (whose error names `agend-terminal app`). No process-global.
-    #[test]
-    fn daemon_capability_reaches_restart_machinery_not_app() {
-        with_env_and_reset(
-            &[("AGEND_RESTART_HANDOFF", "0")],
-            &[
-                "AGEND_WRAPPED",
-                "AGEND_SUPERVISED",
-                "INVOCATION_ID",
-                "XPC_SERVICE_NAME",
-            ],
-            || {
-                let tmp = unique_tmp("daemon-cap");
-                let response = handle_restart_daemon(
-                    &tmp,
-                    Some(crate::api::RestartCapability::Daemon),
-                    None,
-                    None,
-                );
-                assert_eq!(
-                    response["ok"], false,
-                    "unsupervised daemon restart fails closed on the supervisor check, got {response}"
-                );
-                let error = response["error"].as_str().unwrap_or("");
-                assert!(
-                    error.contains("supervisor"),
-                    "Daemon capability must REACH the daemon strategy (supervisor-required error) — got {error:?}"
-                );
-                assert!(
-                    !error.contains("agend-terminal app"),
-                    "Daemon capability must NOT return the app fail-close message — got {error:?}"
-                );
-                let _ = std::fs::remove_dir_all(&tmp);
-            },
-        );
-    }
-
-    /// #2453 Stage R1: an INJECTED `RestartCapability::App` fail-closes with the
-    /// app-specific message and MUST NOT set `RESTART_PENDING` or write the
-    /// restart-requested marker (the successor-spawn path lives ONLY in the
-    /// Daemon strategy, unreachable from this arm).
-    #[test]
-    fn app_capability_fails_closed_leaves_restart_pending_false() {
-        with_env_and_reset(&[("AGEND_RESTART_HANDOFF", "1")], &[], || {
-            let tmp = unique_tmp("app-cap");
-            let response =
-                handle_restart_daemon(&tmp, Some(crate::api::RestartCapability::App), None, None);
-            assert_eq!(
-                response["ok"], false,
-                "app-capability restart must fail closed, got {response}"
-            );
-            let error = response["error"].as_str().unwrap_or("");
-            assert!(
-                error.contains("app"),
-                "app fail-close error must name `app` mode (actionable) — got {error:?}"
-            );
-            assert!(
-                !crate::daemon::RESTART_PENDING.load(Ordering::Acquire),
-                "app capability must NOT set RESTART_PENDING"
-            );
-            assert!(
-                !tmp.join("restart-requested").exists(),
-                "app capability must NOT write restart-requested (no successor spawned)"
-            );
-            let _ = std::fs::remove_dir_all(&tmp);
-        });
-    }
-
-    /// #2453 Stage R1: an INJECTED `RestartCapability::Unsupported` (e.g. the
-    /// `verify` host) is default-deny — fail-closed, no `RESTART_PENDING`, and it
-    /// CANNOT reach the daemon strategy. A DISTINCT value/arm from `App` even
-    /// though both currently share a fail-closed helper.
-    #[test]
-    fn unsupported_capability_default_deny_cannot_reach_daemon() {
-        with_env_and_reset(&[("AGEND_RESTART_HANDOFF", "1")], &[], || {
-            let tmp = unique_tmp("unsupported-cap");
-            let response = handle_restart_daemon(
-                &tmp,
-                Some(crate::api::RestartCapability::Unsupported),
-                None,
-                None,
-            );
-            assert_eq!(
-                response["ok"], false,
-                "unsupported-host restart must default-deny, got {response}"
-            );
-            assert!(
-                !crate::daemon::RESTART_PENDING.load(Ordering::Acquire),
-                "unsupported capability must NOT set RESTART_PENDING"
-            );
-            assert!(
-                !tmp.join("restart-requested").exists(),
-                "unsupported capability must NOT write restart-requested"
-            );
-            let _ = std::fs::remove_dir_all(&tmp);
-        });
-    }
-
-    /// #2453 Stage R1: an ABSENT capability (`None` — a standalone bridge call
-    /// that never traversed the api `mcp_tool` ingress, so carries no
-    /// `RuntimeContext`) default-denies exactly like `Unsupported`.
-    #[test]
-    fn absent_capability_none_default_deny() {
-        with_env_and_reset(&[("AGEND_RESTART_HANDOFF", "1")], &[], || {
-            let tmp = unique_tmp("none-cap");
-            let response = handle_restart_daemon(&tmp, None, None, None);
-            assert_eq!(
-                response["ok"], false,
-                "absent capability (None) must default-deny, got {response}"
-            );
-            assert!(
-                !crate::daemon::RESTART_PENDING.load(Ordering::Acquire),
-                "None capability must NOT set RESTART_PENDING"
-            );
-            let _ = std::fs::remove_dir_all(&tmp);
-        });
-    }
-}
-
 // #2453 R2: the App-arm handler tests live in the sibling `restart_tests.rs` (its
 // name contains "test" so the `file_size_invariant` skips it), keeping this handler
 // implementation file under MAX_LOC. `#[path]` + `#[cfg(test)]` so it compiles only
